Fix: 퀴즈 진척도가 정상적으로 계산되지 않던 오류 수정

+++ src/main/kotlin/com/quizit/quiz/dto/response/GetProgressByIdResponse.kt
@@ -1,6 +1,6 @@
 package com.quizit.quiz.dto.response
 
 data class GetProgressByIdResponse(
-    val total: Long,
-    val solved: Long
+    val total: Int,
+    val solved: Int
 )
